M16827: Possible incorrect term "read/write store"
Hello, @dragon119, 

Translator has reported possible source content issue. Judging from the previous segment, "read-only replicas of the read store" may be a misdescription of "read-only replicas of the write store".

Please, help to check my proposed file change into the article and help to merge if you agree with fix. If not, please, let me know either if you would like me to fix it in another way within this PR, if you prefer to fix it in another PR, or if I should close this PR as by-design. In case of using another PR, please, let me know of your PR number, so we can confirm and close this PR. 

Many thanks in advance.