M28935: Possible extra text at the end of sentence
Hello, @iainfoulds,
Translator has reported possible source content issue. 
"This is the end portion of the sentence in  the previous segment. It is not necessary, and must be removed."

Please review and merge the suggested proposed file change to avoid this error. If you make related fix in another PR, then share your PR number, so we can confirm and close this PR.
Many thanks in advance.