CAF ready v1 review PR (#106)
* Non-dependent files

* Acrolinx spelling issue

* Org readiness content

* Acrolinx fixes

* PR Review updates.

* Title case fixes

* Ready section image updates

* remove svg file.

* Merge update part 1

* Temp next steps remove for merge.

* review fixes

* link fix.

* Add remaining parts of best practices section.

* link fixes

* Duplicate H1 fix.

* link fix

* Rename infrastructure best practices

* link fixes

* text tweak

* Skills readiness update.

* Skills readiness TOC updates.

* Acrolinx review edits

* PR review tweaks.

* Update Azure Readiness Guide content.

* Add landing zone index and reorg Ready section TOC

* Roles and TOC tweak.

* Roles tweaks

* TOC locale fix.

* Integrate storage guidance reviewer feedback.

* Storage guidance tweaks.

* CAF governance journey  tweaks.

* Sync Mike's table header fixes.

* Acrolinx fixes

* CAF Migrate section link fixes

* Acrolix fixes

* Add Compute Decisions article.

* Acrolinx fixes

* Update index.md

The first note section:  It mentions starting with a dedicated Cloud Strategy Team, yet it doesn't discuss a cloud strategy team until later on.  My reco is that if you mention starting with this team, then jump right into describing that team instead of talking about the cloud adoption team first.

Cloud adoption team section:  I'm not sure the team is a MECHANISM for onboarding solutions into the cloud.  I think there's a more accurate word, but I'm not sure what it would be.  Does this team actually onboard solutions?

Cloud center of excellence: I know that you're using this phrase in a way that obviates the need to use the word "team", but it just seems odd not to have the word "team" in there when it's there for "cloud governance".

Partner across teams section: Skills don't have relationships or dependencies, so that sentence seems awkward.

* comments I have that can't easily be added such as via Word 

Team structure section:  "This meeting..."  It's not a meeting, which implies people gathering together for one session.  I think you're trying to say something like a gathering of the minds over a period of time, but "meeting" is not the most accurate word to use here.

* comments I have that can't easily be added such as via Word 

First paragraph:  Is it really "the teams surrounding the cloud adoption teams"?  Or is it "the teams represented in the cloud adoption teams"?

Paragraph that starts off "Depending on the desired business outcomes": is the bullet pointed list really a list of skills?  I think it's a list of roles, so I'd change the sentence to read "...the roles needed on this team could include"

Interactions with supporting teams section:  In this sentence ending in "...cultural blockers that require support from both teams", what are the 2 teams you're referring to?  The sentence prior mentions 4 teams.  

Key responsibilities section:  does one really "mature" existing practices and tools?  Existing practices and tools mature, but I'm not sure people mature them.

* comments I have that can't easily be added such as via Word 

Going back to my comment on the initial org alignment page where I noted that "team" should be added at the end of CCoE, I notice that it's present here on this page after "CCoE"...

* additional comments

Cloud guardian section: The 2nd part of this sentence, "A cloud custodian cleans up the messes made by innovative cloud architects, and the two roles have natural friction and opposing objectives. " feels a bit out of place and perhaps would be better placed after the next sentence about what cloud guardians do.  Or, move the entire sentence and add it to the end of the cloud guardian sentence so it reads "A cloud guardian helps keep the cloud safe, so other cloud architects can move more quickly with fewer messes while a cloud custodian cleans up the messes made by innovative cloud architects, sothe two roles have natural friction and opposing objectives."

* Update skills-readiness-path.md

* additional comments

Identify gaps section:  For number 2, in this phrase "In the case where no area is identified as being responsible..." I'm not sure that an area can be responsible.  Usually responsibilities fall to people or groups.

Partner across teams section:  I don't believe skills themselves can have relationships or dependencies.  Are you trying to say that there are multiple roles that might require a particular skill?

* additional comments

The headline of the page doesn't match what's in the left nav.

Cloud strategy team section:
Under Visibility:  This sentence: "...as an active member of the cloud strategy team, they should also be able to view costs per business unit &/or billing unit, to validate showback, chargeback, or other cloud accounting models."  Are they supposed to be able to view costs in order to validate showback, chargeback, etc?  Or are they supposed to view costs AND validate showback, etc.?  If you intend the former, then the comma after "billing unit" should be removed and probably replaced with "in order".

Cloud adoption team section under Accountability:  This sentence is incomplete: "Instead, to educate the team on the goals and how their decisions impacted those goals. "  And then the very last phrase of the paragraph " ...partner with the cloud strategy team to find better decisions."  I don't think decisions are found.  I'd recommend replacing "find" with a more appropriate verb, perhaps "come up with"?

* CAF Org readiness: integrate Angie's edits.

* Update network-decisions.md

Adding vWAN as an architectural construct, updated the use cases for Hub and Spoke.

* additional comments

Headline doesn't match the left nav

Under the Scope section, shouldn't "Cloud operations team" be linked like the other subsections?

* Link fix

* Integrate Jon Ormond's updates.

* Update cloud-center-excellence.md

* Update cloud-center-excellence.md

* Angie feedback batch 1.

* Angie feedbaack batch2

* Reviewer feedback batch3

* Ready considerations section updates.

* Locale link fixes

* Upload net new images

* Add new articles.

* Acrolinx and temp link fixes.

* PR review fixes.

* Add New Organization section

* Add new plan section content.

* Add linked suggested skills.

* PR review updates

* PR review fixes

* Updates to businss-strategy articles

* Remove broken bookmark

* Add next steps

* Updates to business outcomes

* Updates to Digital Estate

* Remove internal note

* Updates to decision guides

* edits

* CAF Governance updates (part 1)

* CAF Governance updates (part 2)

* Move include file to correct folder

* Fix typo

* Address review comments

* Fix typos

* CAF Infrastructure updates

* CAF Migrate updates (part 1)

* Capitalization

* CAF Migrate updates (part 2)

* Acrolinx issues

* Acrolinx fixes

* CAF Operations updates

* CAF Readiness updates

* Address some PR review feedback

* CAF Readiness updates

* CAF updates (misc)

* CAF: File deletions

* CAF image updates

* Add CAF migration landing zone file

* Updated CAF decision guides and considerations

* Fixed CAF links

* CAF updates

* CAF updates

* Fix CAF links

* Fix CAF links

* Fix CAF links

* CAF fixes for blocking issues

* Fix CAF blocking issues

* Fix CAF blocking issues

* Fix CAF blocking issues

* Fix CAF broken links

* CAF server management articles

* Fix CAF link

* Fixes for blocking issues in CAF

* Fix CAF link

* Fix blocking issues in CAF