Remove duplicate content
(this is a redo of https://github.com/MicrosoftDocs/architecture-center/pull/2472 in the private repo)

We have both https://docs.microsoft.com/azure/architecture/reference-architectures/hybrid-networking/hub-spoke and https://docs.microsoft.com/azure/architecture/reference-architectures/hybrid-networking/shared-services, which appear to have the same content presented in two different ways, with the Shared Services version fleshing a little bit more of the shared services out but otherwise not different.  In fact, there's whole paragraphs that appear to be essentially duplicated.  Given that the hub-spoke version has been edited much more recently, and the shared-services one includes `azbb` (which is hastily backed out in a warning, rather than just deleted), I'm suggesting that the winner should be hub-spoke.  If there is a good reason to have both pages, the content needs to be completely revisited to remove redundancy.