Update healthcare-inference.yml
Standardizing the titleSuffix metadata. Mostly so I can request an exemption to the rules that:
1. titleSuffix must be sentence case (all title suffixes in the repo, except this one, are title case)
2. titleSuffix must refer to a product other than Azure. (Azure Architecture Center articles are about different ways of using Azure products and services, often/usually several of them together.)

Azure Example Scenarios is the standard titleSuffix for Architecture Center example scenario articles, so I'm asking to standardize this one, and for these standards (the others are Azure Solution Ideas and Azure Reference Architectures) to be allowed as exemptions for this repo. Thanks.