Update modern-data-warehouse.yml
An updated discussion MDW architectural diagram to reflect a number of updates/changes required to the existing diagram. (1) AAS is not a default requirement, (2) Azure  Databricks  can be used for curation but is not mandatory and it can be used independently, or as part of, an ADF pipeline. Power BI can access Azure Synapse +/- AAS. Underlying MD page needs updating (refers to Azure Blob not ADLS Gen-2), etc.