Update interservice-communication.md
Original text: "When the status of a delivery changes, the Delivery service sends a delivery status event, such as `DeliveryCreated` or `DeliveryCompleted`. Any service can subscribe to these events. In the current design, the Delivery service is the only subscriber".

Based on the diagram it is Delivery History service that subscribes to Delivery service events.