Add Azure Bastion as the jumpbox recommendation
The traditional VM-based jumpbox is no longer recommended; Azure Bastion is recommended.  Some of the other architectures eg https://docs.microsoft.com/en-us/azure/architecture/reference-architectures/dmz/secure-vnet-dmz?tabs=portal are already there.

This PR keeps the information on how to secure a traditional VM jumpbox but adds Azure Bastion as well.  A reasonable alternative would be to remove all references to a traditional jumpbox VM but I didn't know if we wanted to lose that security information here so I kept it for now.