Change title to that in the user story, because...
"Azure Digital Twins Batch Update" is more ambiguous than "Batch integration with Azure Data Factory for Azure Digital Twins". Also, the shorter name has "Azure Digital Twins" as a modifier of "Batch Update" as if the batch update is a type of batch update that is created or performed by Azure Digital Twins. The longer name provides a better understanding of the relationships of the services and includes Azure Data Factory in the name.