Update azure-arc-hybrid-config-content.md
We had potential use cases in here twice. We're in process of moving "Scenario details" like this below the Architecture section, because our research is coming up that the diagram needs to be "above the fold." So I'm keeping the lower "Potential use cases" section, since that's where we're going to move that info anyway.