Added documentation link for #3191
I've added a documentation link for further details on how to create vnet peerings between different subscriptions in the section on resourcegroups explaining the other options. I did not add updated drawings as I feel that would make a dupliucate of information provided in the cloud adoption framework. I added a link to the relevant information in CAF in the related resources.