Update sol-idea-header.md
This article is a solution idea architecture type. We're differentiating it from the full architectures and giving the reader a chance to request or offer info to "upgrade" the solution idea into a full architecture. That's obviously not clear to you (a writer landing on the Include note for the first time), so I clarified that a bit more.