Update aks-upgrade-practices.md
Updated table as I believe it was wrong and also updated text about node pool scale to make it easier to understand. The way it was written it sounded like the node image from the VMSS creation, which could be a long time ago would be used, but you could have done a node image updated after the creation.