Add/restore diagram of ML pipeline
This image used to be in azure-docs-pr/articles/machine-learning/concept-ml-pipelines.md as ./media/concept-ml-pipelines/pipeline-flow.png. It was deleted from concept-ml-pipelines.md sometime in 2020, it seems. It was also copied to various non-Microsoft web pages about machine learning. The history of the file in azure-docs-pr indicates that the image in the repo is older than what's on the web. Also, this version of the image, which I copied from azure-docs-pr, seems to be better quality than the version that's supplied in user story 1981760.