updating routing logic for azure spring apps
At Microsoft Build this year we announced a product name change and Azure Spring Cloud is now Azure Spring Apps. The name change reflects new product capabilities and positioning for all types of Spring apps. In contrast, Azure Spring Apps was considered a niche product only for microservices when we first started. As such I would like to propose a change in the decision tree and route all types of Spring apps to Azure Spring Apps.