Tweaks to zonal web app architecture
This is a great new reference architecture, but it currently paints zones and regions as competing concepts that are tangential to each other. 

These minor tweaks introduce points of integration with mission-critical scenarios where synergies exist, as well as altering the positioning of regions vs zones to highlight they are supporting constructs while maintaining the focus of this solution on that of a single region.