Fix PoliChek hit: Olympic
User Story https://dev.azure.com/mseng/TechnicalContent/_workitems/edit/2001296. 
Although this usage is acceptable, we're trying to remove as many instances as possible to avoid future hits.
The link itself is a historical (2018) MSDN Magazine article in one of the archive repos. There's no redirect file for these articles and renaming/editing this article probably isn't a good idea. So I left the "olympic" word in the filename.