New useful references added
Adding in several references to more detail at relevant sections of the document.

Describe the summary, scope, and intent of this PR:
The intent was to review this page for currency. The guidance is quite tech-independent so is still current, however there are a few key docs that would be useful to the reader, so I've added in the references.