Update functions-app-security.md
OWASP has published a designated document for serverless application security, which should be referenced not the traditional OWASP TOP 10 doc. I've replaced it.

I'm afraid what's currently written is in contrast with OWASP's view. You can check their doc for further details but hereby, I'm quoting their conclusion: 

"All that means that hackers would have to come up with a different approach for attacks, which means
different attack vectors. The application developers will not be able to put a single traditional perimeter
protection and would need to change their way of thinking, as almost none of the mitigations suggested for
traditional systems would fit in the serverless world."