Update non-relational-data-content.md
Added an additional link for further reading because currently the information is a bit scattered, and I aim to seamlessly guide a reader to satisfy his curiosity, if he aims so. 

Also, moved that paragraph to a Note for better clarity. Many people, including my former past, wrongly thought that flexible schema means "store it the way you like to". But that's just wrong.