Update host-name-preservation-content.md
Adding a minor 2 sentence addition to the end of section "Reverse proxy configuration" 

Verbatim update:
"Whether you preserve or override the host name in the reverse proxy, ensure that the backend server is configured to accept the request in your format. Refer to Back-end Configuration for more information."


Goal:
We got a supportability ask for this, 
despite that it is explained already several times that the backend must be configured to recognize the request (eg Azure PaaS must add as a Custom Domain), 
that it is valuable to make this mention in context explicitly as well in the reverse proxy configuration section.