Update real-time-lakehouse-data-processing-content.md
updates made based on Chad's feedback.

Please talk about how this solution should be considered as part of a Microsoft Fabric world. I'm not sure if it's an "alternative" or should be mentioned in some other way, but we're getting a lot of feedback that Synapse articles must include some statements of relevancy for/with Fabric or if this is a solution that doesn't work within Fabric. If this solution would also work in Fabric, please state that clearly in an alternative. If this solution would NOT also work in Fabric, then state that and what features are missing from Fabric that make this not possible.
{{Added a note, i am planning to submit an equivalent using Fabric}}

The "Azure Synapse Link for Azure Cosmos DB" has been deprecated in favor of mirroring in Microsoft Fabric. -- I know this article doesn't doesn't address fabric -- but when it does (from the prior bullet point), this needs to be addressed probably.
{{Removed references and text}}

Please clarify the role that Azure AI Search has in this architecture through the use of an example. It's mentioned that Azure AI Search is an alternative and mainly around partition inflexibility for Cosmos DB. Let's make sure an example is added to help drive home this scenario better.
{{done}}

In reliability, please address availability zone support for the components. Any recommendations to share?
{{done}}

In cost optimization, provide a link to the pricing calculator with a "typical" buildout of this architecture. {{should i create one using my account?}}

For this being a "near-realtime" architecture, the performance efficiency section seems to be relatively light. Is there any additional guidance to help make this as "near" realtime as possible or is this it?{{it depends on the streaming window and also the amount of transformations data has to go through, happy to discuss and add more here}}