Updating based on comments
Making it clearer that modering appplications using this architecture should consider managed services first, rather the VMs. That said, VMs still have a part to play when migrating legacy systems into Azure as a first step, negating a lot of rework, therefore that section remains, albeit with a bit of rewording.

Also updating the n-tier image to have a margin around the edge (better for dark mode).