Revise AKS CI/CD pipeline documentation (#2)
* Revise AKS CI/CD pipeline documentation

Updated the document to clarify the CI/CD pipeline process for Azure Kubernetes Service (AKS) and Azure Pipelines, including detailed steps and components.

* Clarify repository types in PR pipeline description

* Clarify AKS in CI/CD architecture description

* Refine Azure CI/CD documentation for clarity and consistency

* Update AKS CI/CD documentation for clarity

Reorganized and clarified the monitoring and security sections in the AKS CI/CD documentation.

* Revise AKS CI/CD pipeline architecture description

Updated the architecture description for the AKS CI/CD pipeline to clarify the steps and processes involved.

* Add files via upload

* Update and rename aks-cicd-azure-pipelines-content.md to aks-cicd-azure-pipelines.md

* Delete docs/guide/aks/aks-cicd-azure-pipelines.yml

* Delete docs/browse/thumbs/aks-cicd-azure-pipelines-architecture.png

* Revise title and description for AKS CI/CD guide

Updated the title and description for clarity and detail.

* Change CI/CD guide link from .yml to .md

* Update link format for Azure Pipelines documentation

* Update orchestration-content.md

* Update aks-cicd-azure-pipelines.md

* Clarify CI/CD implementation details for AKS

Updated wording for clarity on Azure Pipelines CI/CD implementation.