Update documentation to reflect Microsoft Agent Framework
Replacing "Semantic Kernel" with the newly announced "Microsoft Agent Framework". Note that the "Next Steps" at the bottom of this article still has "semantic-kernel" in a path for [Overview of the Agent architecture](/semantic-kernel/frameworks/agent/agent-architecture) but that was not changed on purpose because I did not want to change paths of external content, but thought it noteworthy to acknowledge here.